% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_signif_and_format.R
\name{add_signif_and_format}
\alias{add_signif_and_format}
\title{Format P value to text}
\usage{
add_signif_and_format(df, p.col, col.to.format = NULL, decimal = 3)
}
\arguments{
\item{df}{The dataframe / tibble}

\item{p.col}{The name of the p value column}

\item{col.to.format}{Optional, the column to format instead of the p.value colomn itself}

\item{decimal}{Optional, the number of decimal places to round the number to}
}
\value{
The dataframe / tibble with the significance star column and the formated p value column
}
\description{
Format the p value column with stars behind.
}
