% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catNcopy.R
\name{catNcopy}
\alias{catNcopy}
\title{Concatenate & Copy}
\usage{
catNcopy(texts, sep = "\\", \\"")
}
\arguments{
\item{texts}{the text vector}

\item{sep}{the separator}
}
\value{
no return, the values are copied to the clipboard
}
\description{
Helper to concatenate string & copy to clipboard
}
