% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_plot.R
\name{export_plot}
\alias{export_plot}
\title{Export an R plot}
\usage{
export_plot(nullObject = NULL, fileName)
}
\arguments{
\item{nullObject}{is to help for pipe usage}

\item{fileName}{name of the output file}
}
\value{
No return
}
\description{
simply export an R plot using the ggexport device
}
