% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_move.R
\name{file_move}
\alias{file_move}
\title{Move and Rename a File}
\usage{
file_move(from, to)
}
\arguments{
\item{from}{The full path of the original file}

\item{to}{The full path of the destination file, i.e., the new FILENAME at the new folder, not the directory}
}
\value{
return True if successful
}
\description{
A wrapper around file.rename, allowing the destination folder to be created if needed
}
