% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mode.R
\name{get_mode}
\alias{get_mode}
\title{Get the Mode of a Vector}
\usage{
get_mode(x)
}
\arguments{
\item{x}{A vector of numeric or character values.}
}
\value{
The mode of the vector `x`, which is the value that appears most frequently.
}
\description{
This function calculates the mode of a vector, i.e., the most frequent value.
}
\examples{
get_mode(c(1, 2, 2, 3, 3, 3))
get_mode(c("apple", "banana", "apple", "apple", "cherry"))

}
