% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notIn.R
\name{\%!in\%}
\alias{\%!in\%}
\title{Operator: Not in a vector}
\usage{
x \%!in\% y
}
\arguments{
\item{x}{the element to test}

\item{y}{the vector in which the element is searched for}
}
\value{
The created formula
}
\description{
Test whether the element in NOT in a vector
}
