% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printing_pipe.R
\name{\%P>\%}
\alias{\%P>\%}
\title{Printing Pipe}
\usage{
lhs \%P>\% rhs
}
\arguments{
\item{lhs}{the object passed by previous function}

\item{rhs}{the function to print}
}
\value{
No return
}
\description{
the pipe prints the output and passes on the previous value
}
