% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_top_item.R
\name{list_top_item}
\alias{list_top_item}
\title{List Top Items from a Vector}
\usage{
list_top_item(vec, num)
}
\arguments{
\item{vec}{A vector of items (can be characters or factors).}

\item{num}{An integer specifying how many top items to return.}
}
\value{
A formatted string of the top `num` most frequent items along with their counts.
}
\description{
This function returns a formatted string of the top `num` most frequent items from a vector.
}
\examples{
list_top_item(c("apple", "banana", "apple", "cherry", "banana", "banana"), 2)

}
