% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_add_row.R
\name{mat_add_row}
\alias{mat_add_row}
\title{Add a row to a matrix}
\usage{
mat_add_row(M)
}
\arguments{
\item{M}{The matrix}
}
\value{
The matrix with one extra row added at the bottom
}
\description{
Add a row to a matrix
}
