% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify.R
\name{notify}
\alias{notify}
\title{Display a Green Notification Message}
\usage{
notify(msg)
}
\arguments{
\item{msg}{A character string containing the message to be displayed.}
}
\value{
No return value. This function simply prints the message to the console.
}
\description{
This function prints a message to the console in green color to notify the user.
}
\examples{
notify("Operation completed successfully.")

}
