% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_formula.R
\name{paste_formula}
\alias{paste_formula}
\title{Create a formula}
\usage{
paste_formula(y, x)
}
\arguments{
\item{y}{the dependent variable Y}

\item{x}{the independent variable X}
}
\value{
The created formula
}
\description{
Help to create Formula
}
