% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_tibble_row.R
\name{print_tibble_row}
\alias{print_tibble_row}
\alias{transpose_tibble_row}
\title{Print tibble row}
\usage{
print_tibble_row(tbl, row = 1)

transpose_tibble_row(tbl, row = 1)
}
\arguments{
\item{tbl}{the table to print}

\item{row}{which row to print}
}
\value{
A new tibble with name-value pairs
}
\description{
Print the a row of a tibble as name-value pairs
}
