% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renv_reinstall_broken_packages.R
\name{reinstall_broken_packages_for_renv}
\alias{reinstall_broken_packages_for_renv}
\title{Reinstall Broken packages using renv}
\usage{
reinstall_broken_packages_for_renv()
}
\value{
No return
}
\description{
Run renv::restore() and renv::hydrate() first, and there might still be broken packages left, run this to install those broken packages.
}
