% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reveal_in_finder.R
\name{reveal_in_finder}
\alias{reveal_in_finder}
\title{Reveal in Finder}
\usage{
reveal_in_finder(path = getwd())
}
\arguments{
\item{path}{reveal the file or directory in finder}
}
\value{
No return
}
\description{
Show the file or the directory in finder
}
