% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_sample.R
\name{safe_sample}
\alias{safe_sample}
\title{Safe Sampling from a Vector}
\usage{
safe_sample(vector, size)
}
\arguments{
\item{vector}{A vector from which to sample.}

\item{size}{An integer specifying the size of the sample.}
}
\value{
A vector of length `size` if `length(vector) >= size`, otherwise returns the entire `vector`.
}
\description{
This function returns a sample of a specified size from a vector, but if the vector is shorter
than the desired size, it returns the entire vector instead of generating an error.
}
\examples{
safe_sample(1:10, 5)
safe_sample(1:3, 5)

}
