% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plotly.R
\name{save_plotly}
\alias{save_plotly}
\title{Save Plotly Plots to HTML}
\usage{
save_plotly(p, filename)
}
\arguments{
\item{p}{the plot to export}

\item{filename}{the file to save to}
}
\value{
no return, the values are copied to the clipboard
}
\description{
The original function saveWidget doesn't work when used in a directory other than the current working directory (from getwd()). Specifically, it doesn't delete the resource folder.
}
