% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_decimal.R
\name{specify_decimal}
\alias{specify_decimal}
\title{A unified rouding function}
\usage{
specify_decimal(x, k = 2, as_string = T)
}
\arguments{
\item{x}{The number}

\item{y}{The number of decimal places}
}
\value{
The converted string
}
\description{
Convert a number to string with specified decimal places
}
