% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_tmp_csv.R
\name{view_tmp_csv}
\alias{view_tmp_csv}
\title{View table as a temporary csv file}
\usage{
view_tmp_csv(tbl)
}
\arguments{
\item{tbl}{A table or data frame}
}
\value{
Invisibly returns the path to the temporary CSV file.
}
\description{
Write a data.frame to a temporary CSV file to be viewed externally in finder.
}
\details{
The function creates a temporary file with a `.csv` extension, writes the provided table to this file, and then opens the directory containing the file in the MacOS finder. This is useful for quickly inspecting tables outside of R.
}
\examples{

library(dplyr)
library(readr)

# Example table
example_tbl <- tibble::tibble(
  x = 1:5,
  y = letters[1:5]
)

# View the table in a temporary CSV file
view_tmp_csv(example_tbl)

}
