% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_S4_object.R
\name{write_object}
\alias{write_object}
\alias{coerce_object_as_list}
\title{Print class object}
\usage{
write_object(obj, filename, show.in.finder = T)

coerce_object_as_list(obj)
}
\arguments{
\item{filename}{The output file name}

\item{show.in.finder}{Whether to show the written file after printing}

\item{object}{An S4 class object}
}
\value{
No return
}
\description{
Print information from a class object to a file
}
