% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_word.R
\name{write_word}
\alias{write_word}
\title{Export table as word}
\usage{
write_word(table, filename)
}
\arguments{
\item{table}{The table to export}

\item{filename}{The output file name}
}
\value{
No return
}
\description{
Export a normal table (tibble or data.frame) as a word document
}
